添加点击网络错误优化提示
