+++ style/root.scss
@@ -5,4 +5,5 @@
 @import './book.scss';
 @import './footer.scss';
 @import './notebook.scss';
+@import './error.scss';
 @import './media-search.scss';
